+++ Tesla/src/Tesla/ImGui/ImGuiLayer.cpp
@@ -2,7 +2,6 @@
 #include "ImGuiLayer.h"
 
 #include <GLFW/glfw3.h>
-#include <glad/glad.h>
 
 #include "imgui.h"
 #include "backends/imgui_impl_opengl3.cpp"
@@ -74,7 +73,7 @@ namespace Tesla {
 	{
 		ImGuiIO& io = ImGui::GetIO();
 		Application& app = Application::Get();
-		io.DisplaySize = ImVec2(app.GetWindow().GetWidth(), app.GetWindow().GetHeight());
+		io.DisplaySize = ImVec2((float)app.GetWindow().GetWidth(), (float)app.GetWindow().GetHeight());
 
 		// Rendering
 		ImGui::Render();
